requirements.txt
langchain_openai
langchain_chroma
python_dotenv
langchain_community
langchain_core
streamlit
langchain_ollama